Introducing data set descriptions. (BREAKING CHANGE)
After loading the data set, twinify will extract a dictionary of dtypes for each column as well as the overall shape.
Custom models model_factory is no longer provided with the data but only with the data description.
Custom models now may provide a load_data function which is used by twinify to load the data set. This should be used to set proper dtypes for columns.